refactor(i18n): point imports at @/i18n/* and remove legacy files
- Replace @/navigation imports with @/i18n/navigation
- Replace @/middleware constant imports with @/i18n/routing
- Delete src/i18n.ts and src/navigation.ts (replaced by i18n/* split)
