+++ client/src/components/ui/home-link.tsx
@@ -2,7 +2,7 @@
 
 import { useTranslations } from "@/i18n";
 import { cn } from "@/lib/utils";
-import { Link } from "@/navigation";
+import { Link } from "@/i18n/navigation";
 import Image from "next/image";
 
 type HomeLinkProps = {
