--- client/src/navigation.ts
@@ -1,10 +1,0 @@
-import { LOCALES } from "@/middleware";
-import { createSharedPathnamesNavigation } from "next-intl/navigation";
-
-export const locales = LOCALES;
-export const localePrefix = "always"; // Default
-
-export const { Link, redirect, usePathname, useRouter } = createSharedPathnamesNavigation({
-  locales,
-  localePrefix,
-});
